Update to v1.4.0-beta.3
### Added
 Toolbar with different tools including: drag, fire, drop, rope, and delete. Not all are fully working at the moment.
* Currently finished tools: fire, drop, delete
* SDL Licenses.
* Credits to Docs.
* More Doxygen stuff.
* Debug message that prints ball mass.

### Changed
* Moved various items out of main.cpp into core.cpp to avoid main becoming too large.
