+++ src/global.h
@@ -3,6 +3,12 @@
 /*****************************************************************************/
 //This file will hold the global references that are need in one place
 /*****************************************************************************/
+/////////////////////////////////////////////////
+/// @file global.h
+/// @brief This file holds all of the references to variables declared in the
+///        Global namespace. See Main.cpp for more details
+/////////////////////////////////////////////////
+/*****************************************************************************/
 class clsConfig;
 /*****************************************************************************/
 //global typedefs
@@ -15,10 +21,22 @@ namespace cannonballs {
   extern uint intCannonBallNum;
 } //end namespace cannonballs
 
+/////////////////////////////////////////////////
+/// @defgroup GlobalValues Values held globally
+/// @brief Things that will need to be referenced by many
+///        different functions.
+/// @{
+/////////////////////////////////////////////////
+
 namespace global {
   extern const bool blnDebugMode;
   extern clsConfig config;
   namespace physics {
+  /////////////////////////////////////////////////
+  /// @defgroup ChangeablePhysics Defined Physics values
+  /// @brief Physics values for how the program works.
+  /// @{
+  /////////////////////////////////////////////////
     extern const float kBallDensity;
     extern const float kGravity;
     extern const float kDragCofficient;
@@ -28,18 +46,28 @@ namespace global {
     extern const float kCoefficientRestitution;
     extern uchar collisionmethod;
     extern const double kAirDynViscosity;
-  }
+  /// @}
+  } // end physics namespace
 
   namespace equations {
+    /////////////////////////////////////////////////
+    /// @defgroup EquationValues
+    /// @brief Values used for certain equations
+    /// @{
+    /////////////////////////////////////////////////
     extern const float kVelocityScalar;
     extern const float kTimeSizeRatio;
     extern const float kMassAlphaRatio;
     extern const float kMassAlphaOffset;
     extern const float kAlphaMinimum;
-  }
-}
+    /// @}
+  } // end equations namespace
+} // end global namespace
+/// @}
 /*****************************************************************************/
-/** The method of the collision (I just wanted to play around with some options.) */
+/** @enum Collisions
+  * The method of the collision (I just wanted to play around with some options.)
+  */
 enum Collisions {
   CollideElastic = 0, /**< This is normal collision in a perfect world without
                            friction, they hit and bounce off, no energy is lost. */
@@ -50,7 +78,9 @@ enum Collisions {
   CollideNone /**< No collisions, balls pass through each other. */
 };
 /*****************************************************************************/
-/** The current tool being used in the toolbox. */
+/** @enum Tools
+  * The current tool being used in the toolbox.
+  */
 enum Tools {
   ToolFire = 0, /**< The basic firing tool. */
   ToolDrop,     /**< Drops a ball without velocity. */
