+++ src/image_ball.xpm
@@ -1,15 +1,30 @@
 /* XPM */
 static char * image_ball_xpm[] = {
-"10 10 2 1",
-"0	g #000000",
-"1	g #FFFFFF",
-"0001111000",
-"0011111100",
-"0111111110",
-"1111111111",
-"1111111111",
-"1111111111",
-"1111111111",
-"0111111110",
-"0011111100",
-"0001111000"};
+"25 25 2 1",
+" 	c None",
+"@	c #FFFFFF",
+"       @@@@@@@@@@@       ",
+"      @@@@@@@@@@@@@      ",
+"    @@@@@@@@@@@@@@@@@    ",
+"    @@@@@@@@@@@@@@@@@    ",
+"  @@@@@@@@@@@@@@@@@@@@@  ",
+"  @@@@@@@@@@@@@@@@@@@@@  ",
+" @@@@@@@@@@@@@@@@@@@@@@@ ",
+"@@@@@@@@@@@@@@@@@@@@@@@@@",
+"@@@@@@@@@@@@@@@@@@@@@@@@@",
+"@@@@@@@@@@@@@@@@@@@@@@@@@",
+"@@@@@@@@@@@@@@@@@@@@@@@@@",
+"@@@@@@@@@@@@@@@@@@@@@@@@@",
+"@@@@@@@@@@@@@@@@@@@@@@@@@",
+"@@@@@@@@@@@@@@@@@@@@@@@@@",
+"@@@@@@@@@@@@@@@@@@@@@@@@@",
+"@@@@@@@@@@@@@@@@@@@@@@@@@",
+"@@@@@@@@@@@@@@@@@@@@@@@@@",
+"@@@@@@@@@@@@@@@@@@@@@@@@@",
+" @@@@@@@@@@@@@@@@@@@@@@@ ",
+"  @@@@@@@@@@@@@@@@@@@@@  ",
+"  @@@@@@@@@@@@@@@@@@@@@  ",
+"    @@@@@@@@@@@@@@@@@    ",
+"    @@@@@@@@@@@@@@@@@    ",
+"      @@@@@@@@@@@@@      ",
+"       @@@@@@@@@@@       "};
