+++ src/main.cpp
@@ -1,18 +1,20 @@
 /*****************************************************************************/
+/////////////////////////////////////////////////
+/// @file       Main.cpp
+/// @brief      Holds all the main functions.
+/// @author     GamerMan7799
+/// @author     xPUREx
+/// @version    1.4.0-beta.3
+/// @date       2018
+/// @copyright  Public Domain Unlicense.
+/////////////////////////////////////////////////
+/*****************************************************************************/
 //General Todos
 /** BUG (GamerMan7799#1#): The balls will become stuck together for seemingly random reasons */
 /** BUG (GamerMan7799#3#): Balls' velocity becomes NaN quite often */
 /** TODO (GamerMan7799#9#): Allow setting of some Physics Values in Config */
-/*****************************************************************************/
-#include "version.h"
-#include "toolbar.h"
-#include "cannonball.h"
-#include "tick.h"
-#include "rope.h"
-/*****************************************************************************/
-#include <vector>
-typedef std::vector<clsCannonball> VectorCannon;
-typedef std::vector<clsRope> VectorRope;
+/** TODO (GamerMan7799#5#): Make balls bigger? */
+/** TODO (GamerMan7799#9#): Set Max/Min values for mass? */
 /*****************************************************************************/
 //#define DEFINED_USE_R2_VEL_MODDER
 /**< If this is defined, then program will use unrealistic method that will
@@ -25,27 +27,8 @@ typedef std::vector<clsRope> VectorRope;
      colliding balls. The forces are along the same direction as the velocities
      are. It doesn't make sense in the real world.*/
 /*****************************************************************************/
-/** Holds functions that deal with the balls colliding, and creating new balls
-    that are currently in main.cpp. I wanted to get them out of the global scope. */
-namespace cannonballs {
-  void addNew(LOC, LOC, double);
-  void clean_up(void);
-  void checkCollisons(uint);
-  bool checkOverlap(BOX, BOX);
-  void doCollide(uint, uint);
-  uint intCannonBallNum = 0;
-  VectorCannon balls;
-  VectorRope ropes;
-}
-/*****************************************************************************/
-/** Holds functions some core functions for the main */
-namespace core {
-  void fireRandom(void);
-  void handleEvent(void);
-  //clsToolbar toolbar;
-  //clsTick tick;
-  //clsScreen cannonwindow;
-}
+#include "version.h"
+#include "core.h"
 /*****************************************************************************/
 namespace global {
   /** Holds if build is in debug mode, this can happen if
@@ -87,14 +70,14 @@ namespace global {
                                                 meters for the ball */
     const float kMassAlphaRatio = 41.166; /**< The ratio between mass and its
                                                 alpha (transparent) value. \n
-                                                The equation used is Global::Equations::fMassAlphaRatio
+                                                The equation used is Global::Equations::kMassAlphaRatio
                                                 * ln( clsCannonball::props.mass ) +
-                                                Global::Equations::fMassAlphaOffset */
+                                                Global::Equations::kMassAlphaOffset */
     const float kMassAlphaOffset = 130.84; /**< The offset for the equation of
                                                 mass to alpha value \n
-                                                The equation used is Global::Equations::fMassAlphaRatio
+                                                The equation used is Global::Equations::kMassAlphaRatio
                                                 * ln( clsCannonball::props.mass ) +
-                                                Global::Equations::fMassAlphaOffset */
+                                                Global::Equations::kMassAlphaOffset */
     const float kAlphaMinimum = 50; /**< The lowest that the alpha value is allowed to be,
                                          for reference 0 is completely transparent and 255 is completely opaque */
   } //end namespace Equations
@@ -115,90 +98,30 @@ int main(int argc, char *argv[]) {
     fclose(logfile);
   } //end if logging
 
-  clsScreen cannonwindow;
-  clsTick tick;
-  clsToolbar toolbar;
-
   if (global::blnDebugMode) { printf("OS: %s\n",global::config.values.OperatingSystem); }
-  if ( !cannonwindow.getSDLStarted() ) { return 1; } //exit program if there was an error
+  if ( !core::cannonwindow.getSDLStarted() ) { return 1; } //exit program if there was an error
 
   bool quit = false, holding = false, random_fire = false;
   double tempdeltat;
   int fire_tick  = 5000;
-  LOC oldmouse, currentmouse;
   SDL_Event event;
 
   uint ticks_since_clean = 0;
+  char event_return;
 
   do {
-    cannonwindow.clearscreen(); //Clear the screen so new stuff can be drawn
+    core::cannonwindow.clearscreen(); //Clear the screen so new stuff can be drawn
 
     //poll events
-    if (SDL_PollEvent( &event ) != 0) {
-      if ( event.type == SDL_QUIT ) { quit = true; }
-      else if ( event.type == SDL_MOUSEBUTTONDOWN ) {
-        holding = true;
-        tick.startHolding();
-        SDL_GetMouseState(&oldmouse.x, &oldmouse.y);
-        currentmouse = oldmouse;
-      } else if ( event.type == SDL_MOUSEMOTION && holding ) {
-        //Draw the line
-        SDL_GetMouseState(&currentmouse.x, &currentmouse.y );
-      } else if ( event.type == SDL_MOUSEBUTTONUP ) {
-        holding = false;
-        cannonballs::addNew(currentmouse, oldmouse, tick.stopHolding() );
-      } else if ( event.type == SDL_KEYDOWN ) {
-        switch ( event.key.keysym.sym ) {
-        case SDLK_k:
-          //kill all the balls
-          for (int i = 0; i < cannonballs::balls.size(); ++i)
-            { cannonballs::balls[i].blnstarted_ = false;  }
-          break;
-        case SDLK_q:
-        case SDLK_ESCAPE:
-          //quit
-          quit = true;
-          break;
-        case SDLK_r:
-          //stop all motion of balls
-          for (int i = 0; i < cannonballs::balls.size(); ++i)
-            { cannonballs::balls[i].setVelocity({0.0,0.0}); }
-          break;
-        case SDLK_f:
-          if(global::blnDebugMode) { printf("Random fire triggered\n"); }
-          random_fire = !random_fire;
-          break;
-        case SDLK_t:
-          if(global::blnDebugMode) {printf("Toolbox toggled.\n");}
-          toolbar.toggleToolbox();
-          break;
-        case SDLK_RIGHT:
-          toolbar.incrementTool(1);
-          break;
-        case SDLK_LEFT:
-          toolbar.incrementTool(-1);
-          break;
-        case SDLK_1:
-          toolbar.setTool(ToolFire);
-          break;
-        case SDLK_2:
-          toolbar.setTool(ToolDrop);
-          break;
-        case SDLK_3:
-          toolbar.setTool(ToolRope);
-          break;
-        case SDLK_4:
-          toolbar.setTool(ToolDele);
-          break;
-        case SDLK_5:
-          toolbar.setTool(ToolDrag);
-          break;
-        } //end switch key
-      } //end if event
-    } //end if event
-
-    if (holding && toolbar.getTool() == ToolFire)
-      { cannonwindow.drawline(currentmouse, oldmouse); }
+    if (SDL_PollEvent( &event ) != 0) { event_return = core::handleEvent( &event ); }
+    switch (event_return) {
+    case 'q':
+      quit = true;
+      break;
+    case 'f':
+      random_fire = !(random_fire);
+      break;
+    }
 
     if (random_fire) {
       fire_tick++;
@@ -209,7 +132,7 @@ int main(int argc, char *argv[]) {
     }
 
     //Update every ball
-    tempdeltat = tick.getTimeDifference();
+    tempdeltat = core::tick.getTimeDifference();
     for (int i = 0; i < cannonballs::balls.size(); ++i) {
       //Loop through each cannonball
       if (cannonballs::balls[i].blnstarted_) {
@@ -221,337 +144,21 @@ int main(int argc, char *argv[]) {
       } //end if started
     } //end for loop
 
-    toolbar.show();
-    cannonwindow.update(); //Update the screen
+   if (core::holding && core::toolbar.getTool() == ToolFire)
+    { core::cannonwindow.drawline(core::currentmouse, core::oldmouse); }
+
+
+    core::toolbar.show();
+    core::cannonwindow.update(); //Update the screen
     ticks_since_clean++;
     // run a clean up on dead balls if past a certain number
     if (ticks_since_clean >= 1000) {
       cannonballs::clean_up();
       ticks_since_clean = 0;
     }
-
   } while (!quit); //keep looping until we get a quit
 
-  // clear the cannonball vector
-  //cannonballs::balls.clear();
-  //cannonballs::balls.shrink_to_fit();
   cannonballs::balls = VectorCannon();
-
-  //cannonwindow.~clsScreen();
-  //tick.~clsTick();
-  //toolbar.~clsToolbar();
   return 0;
-}
-/*****************************************************************************/
-void cannonballs::addNew(LOC mouseC, LOC mouseO, double HoldTime ) {
-  /////////////////////////////////////////////////
-  /// @brief Will add a new cannonball based on the mouse start and mouse end.
-  ///        Mass of the ball will vary based on Hold Time
-  ///
-  /// @param mouseC = Current Mouse Location in X and Y
-  /// @param mouseO = Old (start of mouse click) Mouse Location in X and Y
-  /// @param HoldTime = Time (in seconds) that the mouse button was held down for
-  /// @return void
-  ///
-  /////////////////////////////////////////////////
-
-  //Get location, vel, and angle
-  double fire_v;
-  double angle;
-  double radius = (double)global::equations::kTimeSizeRatio * sqrt(HoldTime);
-
-  fire_v = -1 * sqrt( pow(mouseC.x - mouseO.x, 2) + pow(mouseC.y - mouseO.y, 2) );
-  fire_v *= (double) global::equations::kVelocityScalar;
-
-  //If the mouse if pointing straight up or straight down make the angle 90
-  //Otherwise calculate the angle with atan.
-  if (mouseC.x == mouseO.x) {
-    angle = (M_PI / 2) * ( (mouseC.y > mouseO.y) ? -1.0 : 1.0 );
-  } else {
-    angle = (double) -1.0 * atan( (double)(mouseC.y - mouseO.y) /
-             (double)(mouseC.x - mouseO.x) );
-    angle += (double)( (mouseC.x < mouseO.x) ? M_PI : 0.0 );
-  } //end if x = x
-
-  //mod mouse start, once again because the top left is 0,0 to SDL
-  mouseO.y = global::config.values.uintScreenHeight - mouseO.y;
-  intCannonBallNum++;
-  clsCannonball tempBall;
-  tempBall.setValues(radius, mouseO,fire_v, angle, intCannonBallNum);
-  balls.push_back(tempBall);
-
-  return;
-}
-/*****************************************************************************/
-void cannonballs::checkCollisons(uint j) {
-
-  /////////////////////////////////////////////////
-  /// @brief Checks ball number j is colliding with another ball and then do collisions.
-  ///
-  /// @param  j = the number in the array that ball we are checking is.
-  /// @return void (all changes if they are colliding is handled in this function).
-  ///
-  /////////////////////////////////////////////////
-
-  BOX A, B;
-  A = balls[j].getBOX();
-  if (balls[j].blncheckphysics_) {
-    for (int i = 0; i < balls.size(); ++i) {
-      if (balls[i].blncheckphysics_ && balls[j].blncheckphysics_ &&  i != j) {
-        B = balls[i].getBOX();
-        if ( checkOverlap(A, B) ) {
-          doCollide(j, i);
-          balls[i].blncheckphysics_ = false;
-          balls[j].blncheckphysics_ = false;
-        } // end if overlap
-      } //end if started and not same ball
-    } //end for loop inner
-  } //end if check physics
-}
-/*****************************************************************************/
-bool cannonballs::checkOverlap(BOX A, BOX B) {
-  /////////////////////////////////////////////////
-  /// @brief Checks if two boxes overlap
-  ///
-  /// @param A = Box for ball A
-  /// @param B = Box for ball B
-  /// @return TRUE / FALSE if they overlap and therefore collide
-  ///
-  /////////////////////////////////////////////////
-
-  if( A.right < B.left ){ return false; }
-  if( A.left > B.right ){ return false; }
-  if( A.bottom < B.top ){ return false; }
-  if( A.top > B.bottom ){ return false; }
-
-  return true;
-}
-/*****************************************************************************/
-void cannonballs::doCollide(uint numA, uint numB) {
-  /////////////////////////////////////////////////
-  /// @brief Will calculate the new velocities of two balls that are colliding
-  ///
-  /// @image html Equations.PNG "The equations used to get the resulting velocities."
-  /// @param numA = the number in the array ball A is
-  /// @param numB = the number in the array ball B is
-  /// @return void (everything is handled inside the function)
-  ///
-  /////////////////////////////////////////////////
-
-  dblXY Avel, Bvel;
-  PP Aprops, Bprops;
-  double aspin, bspin;
-
-
-  Avel = balls[numA].getVelocity();
-  Bvel = balls[numB].getVelocity();
-  Aprops = balls[numA].getPhysicalProps();
-  Bprops = balls[numB].getPhysicalProps();
-
-  aspin = balls[numA].getSpin();
-  bspin = balls[numB].getSpin();
-
-#ifdef DEFINED_USE_R2_VEL_MODDER
-  //This part here has no actual basis on real life,
-  //it is just my attempt at preventing the cannonballs from sticking together
-  LOC CenterA, CenterB, DeltaCenters;
-  double VelModder;
-  CenterA = balls[numA].getplace();
-  CenterB = balls[numB].getplace();
-
-  DeltaCenters.x = abs(CenterA.x - CenterB.x);
-  DeltaCenters.y = abs(CenterA.y - CenterB.y);
-  //Since it is r^2 and the sqrt of this give us r, we just drop the sqrt part to save time
-  VelModder = sqrt(pow((double)DeltaCenters.x,2) + pow((double)DeltaCenters.y,2));
-  VelModder = sqrt(200.0) / VelModder; //200 is the highest value possible
-  if (VelModder < 1.0) {VelModder = 1.0;}
-
-  Avel.x *= (double) VelModder;
-  Avel.y *= (double) VelModder;
-  Bvel.x *= (double) VelModder;
-  Bvel.y *= (double) VelModder;
-  //End of non real stuff
-#endif
-
-  double Aangle, Bangle, ContactAngle;
-  double Atotal_v, Btotal_v;
-  double spin_energy, aspin_energy, bspin_energy;
-
-  dblXY TotalAMomentum, TotalBMomentum;
-
-  // Negative energy is not realistic, but this is just to keep the direction
-  // of the spin
-  aspin_energy = copysign (0.5 * Aprops.interia * pow(aspin,2.0), aspin);
-  bspin_energy = copysign (0.5 * Bprops.interia * pow(bspin,2.0), bspin);
-  spin_energy = aspin_energy + bspin_energy;
-
-  if ( global::physics::collisionmethod != CollidePerfectInelastic ) {
-    //The equations for Perfect Inelastic is much simpler than if not
-    //so I am handling those a bit differently to speed up that method.
-
-    //Equations used can be found and explained here:
-    //https://en.wikipedia.org/wiki/Elastic_collision
-    Atotal_v = sqrt( pow(Avel.x,2) + pow(Avel.y,2) );
-    Btotal_v = sqrt( pow(Bvel.x,2) + pow(Bvel.y,2) );
-
-    //get the angle for both A and B
-    if (Avel.x != 0.0) { Aangle = atan(Avel.y/Avel.x); }
-    else { Aangle = (Avel.y >= 0.0) ? M_PI / 2 : -M_PI / 2; }
-
-    if (Bvel.x != 0.0) { Bangle = atan(Bvel.y/Bvel.x); }
-    else { Bangle = (Bvel.y >= 0.0) ? M_PI / 2 : -M_PI / 2; }
-
-    //Adjust the angle to be the right one.
-    //Since atan will only yield a number between -PI/2 and PI/2 we
-    //have to adjust it if xvel is negative.
-    Aangle += Avel.x < 0.0 ? M_PI : 0;
-    Bangle += Bvel.x < 0.0 ? M_PI : 0;
-    //The contact angle has to be the difference between the two angles but
-    //since sometimes one or the other is negative, we'll use abs to ensure the right number
-    ContactAngle = abs ( abs(Aangle) - abs(Bangle) );
-
-    TotalAMomentum.x = Atotal_v * cos(Aangle - ContactAngle) *
-                      (Aprops.mass - Bprops.mass);
-    TotalAMomentum.x += 2.0 * Bprops.mass * Btotal_v *
-                      cos(Bangle - ContactAngle);
-    TotalAMomentum.x /= (Aprops.mass + Bprops.mass);
-    //x and y formulas are the same until this point
-    TotalAMomentum.y = TotalAMomentum.x;
-    TotalAMomentum.x *= cos(ContactAngle);
-    TotalAMomentum.y *= sin(ContactAngle);
-    TotalAMomentum.x += Atotal_v * sin(Aangle - ContactAngle) *
-                        cos(ContactAngle + (M_PI / 2) );
-    TotalAMomentum.y += Atotal_v * sin(Aangle - ContactAngle) *
-                        sin(ContactAngle + (M_PI / 2) );
-
-    //Now do B
-    TotalBMomentum.x = Btotal_v * cos(Bangle - ContactAngle) *
-                       (Bprops.mass - Aprops.mass);
-    TotalBMomentum.x += 2.0 * Aprops.mass * Atotal_v *
-                        cos(Aangle - ContactAngle);
-    TotalBMomentum.x /= (Aprops.mass + Bprops.mass);
-    //x and y formulas are the same until this point
-    TotalBMomentum.y = TotalBMomentum.x;
-    TotalBMomentum.x *= cos(ContactAngle);
-    TotalBMomentum.y *= sin(ContactAngle);
-    TotalBMomentum.x += Btotal_v * sin(Bangle - ContactAngle) *
-                        cos(ContactAngle + (M_PI / 2) );
-    TotalBMomentum.y += Btotal_v * sin(Bangle - ContactAngle) *
-                        sin(ContactAngle + (M_PI / 2) );
-  } else {
-    TotalAMomentum.x = Aprops.mass * Avel.x + Bprops.mass * Bvel.x;
-    TotalAMomentum.y = Aprops.mass * Avel.y + Bprops.mass * Bvel.y;
-  } //end if Perfect Inelastic or not
-
-  switch (global::physics::collisionmethod) {
-  //figure out what to do based on the collision method
-  case CollidePerfectInelastic:
-    //The balls collide and stick together
-    //Get the new mass.
-    Aprops.mass += Bprops.mass;
-    //Now we have to calculate the new radius, volume, and area so that drag will reflect the new size
-    //we are of course assuming that density stays the same.
-    Aprops.volume = (Aprops.mass / (double)global::physics::kBallDensity);
-    //cbrt = cube root
-    Aprops.radius = cbrt( (double) (3.0*Aprops.volume) / (double) (4.0*M_PI) );
-    Aprops.area = (double) (2.0 * M_PI * pow(Aprops.radius, 2) );
-    Aprops.interia = (double) (2*Aprops.mass* pow(Aprops.radius,2.0) /5.0);
-    //Now calculate the new velocity
-    Avel.x = TotalAMomentum.x / Aprops.mass;
-    Avel.y = TotalAMomentum.y / Aprops.mass;
-    aspin = copysign( sqrt( 2.0 * abs(spin_energy) / Aprops.interia ), spin_energy);
-    //now "kill" cannonball B and update ball A
-    balls[numB].blnstarted_ = false;
-    balls[numA].setPhysicalProps(Aprops);
-    balls[numA].setVelocity(Avel);
-    balls[numA].setSpin(aspin);
-    break;
-  case CollideInelastic:
-    //uses the same equations as below but some energy is lost.
-    TotalAMomentum.x *= (double)global::physics::kCoefficientRestitution;
-    TotalAMomentum.y *= (double)global::physics::kCoefficientRestitution;
-    TotalBMomentum.x *= (double)global::physics::kCoefficientRestitution;
-    TotalBMomentum.y *= (double)global::physics::kCoefficientRestitution;
-    spin_energy *= (double)global::physics::kCoefficientRestitution;
-
-  case CollideElastic:
-    aspin = copysign( sqrt( abs(spin_energy) / Aprops.interia ), spin_energy);
-    bspin = copysign( sqrt( abs(spin_energy) / Bprops.interia ), spin_energy);
-
-    //The balls collide and bounce away from each other
-
-    //All of the heavy lifting is handled above.
-    balls[numA].setVelocity(TotalAMomentum);
-    balls[numB].setVelocity(TotalBMomentum);
-    balls[numA].setSpin(aspin);
-    balls[numB].setSpin(bspin);
-
-#ifdef DEFINED_COLLISION_NORMAL_FORCE
-    if (TotalAMomentum.x != 0.0) { Aangle = atan(TotalAMomentum.y/TotalAMomentum.x); }
-    else { Aangle = (TotalAMomentum.y >= 0.0) ? M_PI / 2 : -M_PI / 2; }
-    if (TotalBMomentum.x != 0.0) { Bangle = atan(TotalBMomentum.y/TotalBMomentum.x); }
-    else { Bangle = (TotalBMomentum.y >= 0.0) ? M_PI / 2 : -M_PI / 2; }
-    Aangle += TotalAMomentum.x < 0.0 ? M_PI : 0;
-    Bangle += TotalBMomentum.x < 0.0 ? M_PI : 0;
-
-    double normal_force;
-    dblXY forces;
-    normal_force = -1 * Aprops.mass * global::physics::kGravity;
-    forces.x = normal_force * cos (Aangle);
-    forces.y = normal_force * sin (Aangle);
-    balls[numA].addForce(forces);
-
-    normal_force = -1 * Bprops.mass * global::physics::kGravity;
-    forces.x = normal_force * cos (Bangle);
-    forces.y = normal_force * sin (Bangle);
-    balls[numB].addForce(forces);
-#endif
-
-    break;
-  default: //the catch all and CollideNone
-    //Nothing Happens!
-    break;
-  } //end switch collide method
-} //end do Collide
-/*****************************************************************************/
-void cannonballs::clean_up() {
-  /////////////////////////////////////////////////
-  /// @brief Removes any "dead" balls from the balls vector and shrinks it
-  ///        to reduce memory usage
-  /////////////////////////////////////////////////
-  int new_cannon_num = 0; // keeps track of the number of valid balls found
-
-  for(int i = 0; i < intCannonBallNum; ++i) {
-    if ( !(balls[i].blnstarted_) ) {
-      balls.erase(balls.begin()+i);
-    } else { new_cannon_num++; }
-  }
-
-  if (global::blnDebugMode) {
-    printf("Clean up has been run.\n");
-    printf("%3i entities have been cleared\n",intCannonBallNum-new_cannon_num);
-    printf("%3i entities remain\n",new_cannon_num);
-  }
-  intCannonBallNum = new_cannon_num;
-  balls.shrink_to_fit();
-}
-/*****************************************************************************/
-void core::fireRandom() {
-  /////////////////////////////////////////////////
-  /// @brief Causes cannonballs to be fired randomly
-  /////////////////////////////////////////////////
-  Configures cnfg = global::config.values;
-  LOC mouseo, mousec;
-  // time delay can be anywhere from 0.25 to 10 seconds
-  double time_delay = ((rand() % (10000-250) + 250) / 1000);
-
-  mouseo.x = rand() % cnfg.uintScreenWidth;
-  mouseo.y = rand() % cnfg.uintScreenHeight;
-
-  mousec.x = rand() % cnfg.uintScreenWidth;
-  mousec.y = rand() % cnfg.uintScreenHeight;
-
-  cannonballs::addNew(mousec, mouseo, time_delay);
-
 }
 /*****************************************************************************/
