+++ src/rope.cpp
@@ -1,11 +1,12 @@
 /*****************************************************************************/
 #include "rope.h"
 /*****************************************************************************/
+/////////////////////////////////////////////////
+/// @file rope.cpp
+/// @brief Holds all of the functions for the rope
+/////////////////////////////////////////////////
+/*****************************************************************************/
 clsRope::clsRope() {
     //ctor
-}
-/*****************************************************************************/
-clsRope::~clsRope() {
-    //dtor
 }
 /*****************************************************************************/
