+++ src/screen.cpp
@@ -1,6 +1,11 @@
 /*****************************************************************************/
 #include "screen.h"
 /*****************************************************************************/
+/////////////////////////////////////////////////
+/// @file screen.cpp
+/// @brief Holds all of the functions for the SDL screen
+/////////////////////////////////////////////////
+/*****************************************************************************/
 //Include the xpm files which present the images.
 #include "image_ball.xpm"
 #include "image_pixel.xpm"
