+++ src/tick.cpp
@@ -4,6 +4,11 @@
 #include <cstdio>
 #include <cstdlib>
 /*****************************************************************************/
+/////////////////////////////////////////////////
+/// @file tick.cpp
+/// @brief Holds functions for timey wimey stuff
+/////////////////////////////////////////////////
+/*****************************************************************************/
 clsTick::clsTick() {
   /////////////////////////////////////////////////
   /// @brief Default constructor, just sets lngTimerStart to now
