+++ src/tick.h
@@ -4,6 +4,11 @@
 #include <time.h>
 #include <cmath>
 /*****************************************************************************/
+/////////////////////////////////////////////////
+/// @class clsTick tick.h "source/tick.h"
+/// @brief This class will hold functions related to time.
+/////////////////////////////////////////////////
+/*****************************************************************************/
 class clsTick {
   public:
     clsTick();
