+++ src/toolbar.h
@@ -3,10 +3,15 @@
 /*****************************************************************************/
 #include "screen.h"
 /*****************************************************************************/
+/////////////////////////////////////////////////
+/// @class clsToolbar toolbar.h "source/toolbar.h"
+/// @brief This class is used to handle showing the toolbar. The actual functions
+///        of the tools are handled in core.cpp instead.
+/////////////////////////////////////////////////
+/*****************************************************************************/
 class clsToolbar {
     public:
       clsToolbar();
-      ~clsToolbar();
       void show(void);
       void setTool(uchar);
       uchar getTool(void);
