+++ src/version.h
@@ -1,54 +1,64 @@
 #ifndef __VERSION_HEADER__
 #define __VERSION_HEADER__
 /*****************************************************************************/
-	//Date Version Types
-	#define DEFINED_VER_DATE "19"
-	/**< Day of the last build */
+/////////////////////////////////////////////////
+/// @file version.h
+/// @brief  This is the header which holds all of the defines related to
+/// version information.
+/////////////////////////////////////////////////
+/*****************************************************************************/
+/////////////////////////////////////////////////
+/// @defgroup VersionInfo Version Information
+/// @brief These are defines hold all the information that is related to program version.
+/// @{
+/////////////////////////////////////////////////
+/*****************************************************************************/
+//Date Version Types
+/** The day of the last build of the program, currently not used. */
+#define DEFINED_VER_DATE "20"
 
-	#define DEFINED_VER_MONTH "03"
-	/**< Month of the last build */
+/** The month of the last build of the program, currently not used. */
+#define DEFINED_VER_MONTH "03"
 
-	#define DEFINED_VER_YEAR "2018"
-	/**< Year of the last build */
+/** The year of the last build of the program, currently not used. */
+#define DEFINED_VER_YEAR "2018"
 
-	#define DEFINED_VER_UBUNTU_VERSION_STYLE  "18.03"
-	/**< Version styling for Ubuntu, consists of YY.MM */
+/** The Ubuntu style of the date of the last build. It is in YY.MM format. */
+#define DEFINED_VER_UBUNTU_VERSION_STYLE  "18.03"
+/*****************************************************************************/
+/** The Major number of the version number */
+#define DEFINED_VER_MAJOR  1
 
+/** The Minor number of the version number */
+#define DEFINED_VER_MINOR  4
 
-	//Software Status
-	#define DEFINED_VER_STATUS  "Beta"
-    /**<
-    Software Status can be the following:
-    | Status | Short | Description |
-    |-------------------|-------|-------------------------------------------------------------------------------------------------------------------------------------------------|
-    | Release | r | The supported release of the program. |
-    | Release Candidate | rc | The next likely release version; usually only a few bugs / todos separate this froma release. It is pretty safe to use. |
-    | Beta | b | This version will compile and / or run but does not work as expected (or crashes often). |
-    | Alpha | a | This version does not work at all; it usually won't compile at all or crashes more times than it works; best not to use versions in this state. |
-    */
+/** The Patch number of the version number */
+#define DEFINED_VER_PATCH  0
+/*****************************************************************************/
+/** The version number in a format used by Boilerplate.rc, It is in the following format
+    [MAJOR].[MINOR].[PATCH].[BUILDNUMBER] */
+#define DEFINED_VER_RC_FILEVERSION 1,4,0,3
 
-	#define DEFINED_VER_STATUS_SHORT  "B"
-	/**< See table for DEFINED_VER_STATUS */
+/** Is the same as DEFINED_VER_RC_FILEVERSION but is a null terminated string */
+#define DEFINED_VER_RC_FILEVERSION_STRING "1, 4, 0, 3\0"
 
-	//Standard Version Type
-	#define DEFINED_VER_MAJOR  1
-	/**< The major version number */
+/** A more specific string of the file version */
+#define DEFINED_VER_FULLVERSION_STRING  "1.4.0-beta.3\0"
 
-	#define DEFINED_VER_MINOR  4
-	/**< The minor version number */
+//Software Status
+/**
+Software Status can be the following:
+| Status | Short | Description |
+|-------------------|-------|-------------------------------------------------------------------------------------------------------------------------------------------------|
+| Release | r | The supported release of the program. |
+| Release Candidate | rc | The next likely release version; usually only a few bugs / todos separate this from a release. It is pretty safe to use. |
+| Beta | b | This version will compile and / or run but does not work as expected (or crashes often). |
+| Alpha | a | This version does not work at all; it usually won't compile at all or crashes more times than it works; best not to use versions in this state. |
+*/
+#define DEFINED_VER_STATUS 	"Beta"
 
-	#define DEFINED_VER_BUILD  0
-	/**< The patch version number */
-
-	#define DEFINED_VER_RC_FILEVERSION 1,4,0,0
-	/**< The number used in the boilerplate.rc file
-        Follows MAJOR,MINOR,PATCH,BUILD where build is the number of times this version has been built*/
-
-	#define DEFINED_VER_RC_FILEVERSION_STRING "1, 4, 0, 0\0"
-	/**< Is the same as DEFINED_VER_RC_FILEVERSION but is a null terminated string */
-
-	#define DEFINED_VER_FULLVERSION_STRING  "1.4.0-Beta.2"
-	/**< The full string of the version number */
-
+/** See the table for DEFINED_VER_STATUS */
+#define DEFINED_VER_STATUS_SHORT  "b"
+/// @}
 /*****************************************************************************/
 #endif //__VERSION_HEADER__
